A Networking project to make connection of multiple floors using hybrid topology.
You are required to setup a network on 7 floors of a building with each floor consisting of 5 computers. Arrange the computers in a hybrid topology. Ensure connectivity among all the devices using switches. Use different networks for each floor. Configure static routing for ensuring connectivity. Use network 192.168.123.0 with subnet mask 255.255.255.0. Configure DNS service in a way that it is accessible to all the devices connected to the network.